+++ CHANGELOG.md
@@ -1,4 +1,7 @@
-# [4.9.7+x]
+# [4.9.8+x]
+* SmartDialog.config.checkExist() adjust to SmartDialog.checkExist()
+
+* # [4.9.7+x]
 * optimize bindWidget, when bindWidget is not null, bindPage will be automatically set to false.
 * Reconstruct the implementation of initialization
 * Show Debug Paint: fix [#109](https://github.com/fluttercandies/flutter_smart_dialog/issues/109), [#115](https://github.com/fluttercandies/flutter_smart_dialog/issues/115)